dataplatform/tests/requirements.txt
hdfs==2.5.7
kafka-python==2.0.2
python-dotenv==0.8.2
requests==2.31.0
urllib3==2.0.4
paho-mqtt==1.6.1
matplotlib==3.1.3
notebook==5.7.8
pandas==1.3.3
requests==2.31.0
pymongo==4.4.1
pyspark==3.3.0
